Split up LockExpression
